+++ newlib/libc/stdlib/wcstoumax.c
@@ -107,14 +107,14 @@ wcstoumax_l(const wchar_t * __restrict nptr,
 			c -= L'a' - 10;
 		else
 			break;
-		if (c >= base)
+		if ((int) c >= base)
 			break;
-		if (any < 0 || acc > cutoff || (acc == cutoff && c > cutlim))
+		if (any < 0 || acc > cutoff || (acc == cutoff && (int) c > cutlim))
 			any = -1;
 		else {
 			any = 1;
 			acc *= base;
-			acc += c;
+			acc += (uintmax_t) c;
 		}
 	}
 	if (any < 0) {
